+++ src/index.ts
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 // eslint-disable-next-line n/shebang
-export * from './command';
+export * from './run';